+++ common/Dtos/dto.h
@@ -20,10 +20,10 @@ public:
     virtual std::string return_line();
     virtual uint8_t orientation();
     virtual bool is_alive();
-    virtual uint8_t x_pos();
-    virtual uint8_t y_pos();
-    virtual uint8_t return_alto();
-    virtual uint8_t return_ancho();
+    virtual uint16_t x_pos();
+    virtual uint16_t y_pos();
+    virtual uint16_t return_alto();
+    virtual uint16_t return_ancho();
     uint8_t return_code();
 };
 #endif
